chore: rename workloadconfig to instrumentation
