+++ pkg/clusteragent/instrumentation/cel.go
@@ -5,7 +5,7 @@
 
 //go:build kubeapiserver
 
-package workloadconfig
+package instrumentation
 
 import (
 	"fmt"
